requirements.txt
python-dotenv
youtube-transcript-api
openai